requirements.txt
# Temporal SDK
temporalio==1.5.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Validation
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging
structlog==23.2.0

# Utilities
python-dotenv==1.0.0
